@@ -1,13 +1,8 @@
-module github.com/justjanne/powerline-go
+module github.com/thefynx/powerline-go
 
-go 1.15
+go 1.23
 
 require (
-	github.com/mattn/go-runewidth v0.0.9
-	github.com/shirou/gopsutil/v3 v3.22.3
-	golang.org/x/sys v0.0.0-20220722155257-8c9f86f7a55f
-	golang.org/x/term v0.0.0-20210927222741-03fcf44c2211
-	golang.org/x/text v0.3.8
-	gopkg.in/ini.v1 v1.66.4
-	gopkg.in/yaml.v2 v2.4.0
+	github.com/mattn/go-runewidth v0.0.16
+	github.com/shirou/gopsutil/v3 v3.24.5
 )
